actions: Split out zerver.actions.bots.
