+++ zerver/tests/test_audit_log.py
@@ -6,6 +6,12 @@ from django.contrib.auth.password_validation import validate_password
 from django.utils.timezone import now as timezone_now
 
 from analytics.models import StreamCount
+from zerver.actions.bots import (
+    do_change_bot_owner,
+    do_change_default_all_public_streams,
+    do_change_default_events_register_stream,
+    do_change_default_sending_stream,
+)
 from zerver.actions.create_user import (
     do_activate_mirror_dummy_user,
     do_create_user,
@@ -31,10 +37,6 @@ from zerver.actions.user_settings import (
 from zerver.actions.users import do_change_user_role, do_deactivate_user
 from zerver.lib.actions import (
     do_add_realm_domain,
-    do_change_bot_owner,
-    do_change_default_all_public_streams,
-    do_change_default_events_register_stream,
-    do_change_default_sending_stream,
     do_change_realm_domain,
     do_deactivate_realm,
     do_reactivate_realm,
